requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
# Note: PyTorch is installed in Dockerfile with GPU/CPU auto-detection
openai-whisper==20231117
tqdm
ollama==0.1.7
openai==1.3.0
anthropic==0.7.0
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
numpy<1.25.0
scipy<1.12.0
scikit-learn<1.4.0
nltk==3.8.1
pydub==0.25.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
Pillow==10.0.1
pathlib2==2.3.7
yt-dlp>=2024.12.23
youtube-dl>=2021.12.17

# Computer Use / Browser Automation Dependencies
pyautogui==0.9.54
opencv-python==4.8.1.78
psutil==5.9.6
python-xlib==0.33
pynput==1.7.6
